// src/types/precheck.ts
export interface PrecheckPayload {
  session: Session; // TOKYO SESSION', 'LONDON SESSION', 'NEW YORK SESSION'
  asset: string; // Records the Asset ticker symbol
  riskTier: RiskTier; // NORMAL', 'REDUCED', 'PRESERVATION'
  acctBalance: number; // saves current account balance for the trade
  plannedRisk: number; // planned risk percentage for given trade
  macroBias: MacroBias; // 'BULLISH', 'BEARISH', 'RANGE'
  macroCurrentPhase: MacroCurrentPhase; // 'ACCUMULATION', 'EXPANSION', 'DISTRIBUTION'
  keyHTFResistanceFrom: number; // lowest price point of the resistance zone
  keyHTFResistanceTo: number; // highest price point of the resistance zone
  keyHTFSupportFrom: number; // lowest price point of the resistance zone
  keyHTFSupportTo: number; // highest price point of the resistance zone
  macroLiquidityDraw: number; // price of different macro Liquidity Draw
  liquidityTargetDirection: LiquidityTargetDirection; // 'ABOVE', 'BELOW', 'RANGE_BOUNDARY'
  bullishMacroBiasInvalidationLevel: number; // price point of the point where the market will shift to a Bullish Bias if reached
  bearishMacroBiasInvalidationLevel: number; // price point of the point where the market will shift to a Bearish Bias if reached
  macroClarityScore: number; // Total calculated score based on the responses in the Macro Section
  midStructureSequence: MidStructureSequence; // 'HH_HL', 'LL_LH', 'RANGE'
  microMacroAlignment: boolean; // returns True/False if the Micro Structure is aligned with the Macro Bias
  recentStructuralEvent: RecentStructuralEvent; // 'BOS', 'CHOCH', 'NONE'
  liquiditySweepObserved: boolean; // returns True/False if a Liquidity Sweep is observed
  midStructureRegime: MidStructureRegime; // 'CLEAN_TREND', 'WEAK_TREND', 'CHOP'
  midStructureRsiState: MidStructureRsiState; // 'TRENDING', 'DIVERGENCE', 'NEUTRAL'
  midStructureEmaState: MidStructureEmaState; // 'STACKED', 'STACKED_EXPANDING', 'STACKED_COMPRESSING', 'COMPRESSED', 'FLAT'
  mid1mADXStateStrength: Mid1mAdxstateStrength; // 'STRONG (>20)', 'WEAK (<20)'
  mid1mADXStateMomentum: Mid1mAdxstateMomentum; // RISING', 'FALLING', 'FLATTENING'
  mid5mADXStateStrength: Mid5mAdxstateStrength; // 'STRONG (>20)', 'WEAK (<20)'
  mid5mADXStateMomentum: Mid5mAdxstateMomentum; // RISING', 'FALLING', 'FLATTENING'
  midStructureScore: number; // Total calculated score based on the responses in the Mid-Structure Section
  primaryStrategy: PrimaryStrategy; // 'TC2', 'TC1', 'TCL', 'CH1', 'SMOG', 'G2', '35A'
  oneSentenceThesis: string; // A single sentence that is well thought out explaining the thesis behind the trade idea
  strategyFitScore: number; // Total calculated score based on the responses in the Strategy Fit Section
  entryType: EntryType; // 'MARKET', 'LIMIT'
  tradeDirection: TradeDirection; // LONG', 'SHORT'
  entryPrice: number; // Trade idea entry price
  stopPrice: number; // trade idea stop loss price
  targetPrice: number; // trade idea target price
  plannedRR: number; // calculates the planned R:R based on the entry, stop, target price, and position size.
  positionSize: number; // calculated position size based on planned risk and entry and stop loss price
  primaryStopBasis: PrimaryStopBasis; // 'STRUCTURAL_INVALIDATION'
  secondaryExitLogic: SecondaryExitLogic; // 'LIQUIDITY_BASED_TRAIL', 'VOLATILITY_BUFFER', 'TIME_BASED_EXIT', 'CONFIRMATION_FAILURE'
  tradeManagementPlan: TradeManagementPlan; // provides which choice of trading management plan this trade idea will take
  partialAtR: number; // at what point do you take partial profit based on a given positive R level
  moveStopToBreakEvenAtR: number; // at what point do you move stop loss to the break even point based on a given positive R level
  executionQualityScore: number; // Total calculated score based on the responses in the Execution Quality Section
}